Projeto/SCEAS/project/requirements.txt
django 
djangorestframework
drf-yasg
firebase-admin